// src/api/generated-schemas/schemas.ts
export const AutoBuildParametersDtoSchema = zod.object({
  multiStepUndoTimeoutMs: zod.number().int().positive().max(72000000),
  multiUndoIncrement: zod.number().int().positive().max(20),
  saveBuild: zod.boolean(),
  forceSaveMetrics: zod.boolean(),
});
// ...
export const ScheduleParametersDtoSchema = zod.object({
  autoBuildParametersDto: zod.lazy(() => AutoBuildParametersDtoSchema),
  costParameters: zod.array(zod.string()),
});